+++ src/engraving/layout/tlayout.cpp
@@ -35,6 +35,8 @@
 #include "../libmscore/arpeggio.h"
 #include "../libmscore/articulation.h"
 
+#include "../libmscore/bagpembell.h"
+
 #include "../libmscore/note.h"
 
 #include "../libmscore/staff.h"
@@ -427,4 +429,74 @@ void TLayout::layout(Articulation* item, LayoutContext&)
     }
 
     item->setbbox(bRect.translated(-0.5 * bRect.width(), 0.0));
+}
+
+void TLayout::layout(BagpipeEmbellishment* item, LayoutContext&)
+{
+    /*
+    if (_embelType == 0 || _embelType == 8 || _embelType == 9) {
+          LOGD("BagpipeEmbellishment::layout st %d", _embelType);
+          }
+     */
+    SymId headsym = SymId::noteheadBlack;
+    SymId flagsym = SymId::flag32ndUp;
+
+    noteList nl = item->getNoteList();
+    BagpipeEmbellishment::BEDrawingDataX dx(headsym, flagsym, item->magS(), item->score()->spatium(), static_cast<int>(nl.size()));
+
+    item->setbbox(RectF());
+    /*
+    if (_embelType == 0 || _embelType == 8 || _embelType == 9) {
+          symMetrics("headsym", headsym);
+          symMetrics("flagsym", flagsym);
+          LOGD("mags %f headw %f headp %f spatium %f xl %f",
+                 dx.mags, dx.headw, dx.headp, dx.spatium, dx.xl);
+          }
+     */
+
+    bool drawFlag = nl.size() == 1;
+
+    // draw the notes including stem, (optional) flag and (optional) ledger line
+    double x = dx.xl;
+    for (int note : nl) {
+        int line = BagpipeEmbellishment::BagpipeNoteInfoList[note].line;
+        BagpipeEmbellishment::BEDrawingDataY dy(line, item->score()->spatium());
+
+        // head
+        item->addbbox(item->score()->engravingFont()->bbox(headsym, dx.mags).translated(PointF(x - dx.lw * .5 - dx.headw, dy.y2)));
+        /*
+        if (_embelType == 0 || _embelType == 8 || _embelType == 9) {
+              printBBox(" notehead", bbox());
+              }
+         */
+
+        // stem
+        // highest top of stems actually used is y1b
+        item->addbbox(RectF(x - dx.lw * .5 - dx.headw, dy.y1b, dx.lw, dy.y2 - dy.y1b));
+        /*
+        if (_embelType == 0 || _embelType == 8 || _embelType == 9) {
+              printBBox(" notehead + stem", bbox());
+              }
+         */
+
+        // flag
+        if (drawFlag) {
+            item->addbbox(item->score()->engravingFont()->bbox(flagsym,
+                                                               dx.mags).translated(PointF(x - dx.lw * .5 + dx.xcorr, dy.y1f + dy.ycorr)));
+            // printBBox(" notehead + stem + flag", bbox());
+        }
+
+        // draw the ledger line for high A
+        if (line == -2) {
+            item->addbbox(RectF(x - dx.headw * 1.5 - dx.lw * .5, dy.y2 - dx.lw * 2, dx.headw * 2, dx.lw));
+            /*
+            if (_embelType == 8) {
+                  printBBox(" notehead + stem + ledger line", bbox());
+                  }
+             */
+        }
+
+        // move x to next note x position
+        x += dx.headp;
+    }
 }
